add: LMDB backend with initial deep node insert
Still very untested, but it *should* be atleast in theory be working.
The main interface for the storage backend is in database/mod.rs, the
underlying datastructure is in store.rs and the model definition is in
datamodel.rs.
Note: The API is very ugly and it would be really nice to make it more
unified and generic, but it isn't right now.